Move fixtures to __fixtures__ and reduce published files
