+++ change/beachball-e8c5006e-16e2-4ae2-900e-9919f8afb87c.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Specify `files` in package.json to ensure only intended files are published",
+  "packageName": "beachball",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
