+++ package.json
@@ -16,6 +16,10 @@
   "engines": {
     "node": ">=12.0.0"
   },
+  "files": [
+    "bin",
+    "lib/!(__*)"
+  ],
   "scripts": {
     "build": "tsc",
     "change": "node ./lib/cli.js change",
