+++ src/__e2e__/bump.test.ts
@@ -1,12 +1,12 @@
 import fs from 'fs-extra';
-import { RepositoryFactory } from '../fixtures/repository';
+import { RepositoryFactory } from '../__fixtures__/repository';
 import { writeChangeFiles } from '../changefile/writeChangeFiles';
 import { git } from 'workspace-tools';
 import { bump } from '../commands/bump';
 import { getPackageInfos } from '../monorepo/getPackageInfos';
 import { BeachballOptions } from '../types/BeachballOptions';
 import { getChangePath } from '../paths';
-import { MonoRepoFactory } from '../fixtures/monorepo';
+import { MonoRepoFactory } from '../__fixtures__/monorepo';
 import path from 'path';
 
 describe('version bumping', () => {
