+++ src/__e2e__/syncE2E.test.ts
@@ -2,8 +2,8 @@ import fs from 'fs-extra';
 import path from 'path';
 import * as tmp from 'tmp';
 import { sync } from '../commands/sync';
-import { Registry } from '../fixtures/registry';
-import { Repository, RepositoryFactory } from '../fixtures/repository';
+import { Registry } from '../__fixtures__/registry';
+import { Repository, RepositoryFactory } from '../__fixtures__/repository';
 import { getPackageInfos } from '../monorepo/getPackageInfos';
 import { infoFromPackageJson } from '../monorepo/infoFromPackageJson';
 import { packagePublish } from '../packageManager/packagePublish';
