+++ src/__fixtures__/registry.ts
