+++ src/__fixtures__/verdaccio.js
